requirements.txt
py-cord[voice]
yt-dlp
PyNaCl
python-dotenv
